refactor(interface): restructuring files
