+++ src/data/index.ts
@@ -1,4 +1,4 @@
-import { TravelBingoGamesData } from './interfaces'
+import { TravelBingoGamesData } from './game.interfaces'
 import testGame from './test-game'
 
 export default () => {
