+++ src/data/types.ts
@@ -1,0 +1,17 @@
+export type Markdown = string
+export type Date = string
+export type Url = string
+export type Icon = string
+export type Color = string
+
+export interface LocalizedString {
+  [lang: string]: string
+}
+
+export interface LocalizedMarkdown {
+  [lang: string]: Markdown
+}
+
+export interface KeyObject {
+  key: string
+}
